server/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
pillow
psycopg2-binary
dj-database-url
django-extensions
qrcode
